requirements.txt
fastapi
pydantic
python-dotenv
openai
pinecone
uvicorn
